chore(deps): Upgrade jol-core to latest
## Description:
To update jol-core some other elements need to be upgraded, particularly
slice. Code has to be updated to work with 64 bit class layout sizes.

+++ presto-common/src/main/java/com/facebook/presto/common/array/ReferenceCountMap.java
@@ -35,7 +35,7 @@ import static java.lang.reflect.Array.getLength;
 public final class ReferenceCountMap
         extends Long2IntOpenHashMap
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(ReferenceCountMap.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(ReferenceCountMap.class).instanceSize();
 
     /**
      * Increments the reference count of an object by 1 and returns the updated reference count
